refactor: one component per file across settings, palette, notebook
Apply the one-component-per-file rule to the remaining multi-component
files on main:

- settings: split controls.tsx into Toggle.tsx and Segmented.tsx and
  point the six settings tabs at the new modules.
- command palette: extract CommandPaletteItem (and its renderHighlight
  helper) into CommandPaletteItem.tsx.
- notebook: extract DataOutputView (and its pickRichest/IMAGE_MIMES
  helpers) into DataOutputView.tsx.

The lazy*/Suspense wrapper files stay grouped per the rule's
code-splitting exception. MarkdownContent's LinkWithWikilink/TaskListLi
are useCallback render adapters bound to props, not standalone
components, so they stay inline.
